+++ cogstat/cogstat_dialogs.py
@@ -28,10 +28,10 @@ def open_data_file():
     return str(QtWidgets.QFileDialog.getOpenFileName(None, _('Open data file'), '',
                                                      '%s (*.ods *.xls *.xlsx *.csv *.txt *.tsv *.dat *.log '
                                                      '*.sav *.zsav *.por *.sas7bdat *.xpt *.dta '
-                                                     '*.rdata *.Rdata *.rds);;'
+                                                     '*.rdata *.Rdata *.rds *.rda);;'
                                                      '%s *.ods *.xls *xlsx;;%s *.csv *.txt *.tsv *.dat *.log;;'
                                                      '%s *.sav *.zsav *.por;;%s *.sas7bdat *.xpt;; %s *.dta;;'
-                                                     '%s *.rdata *.Rdata *.rds' %
+                                                     '%s *.rdata *.Rdata *.rds *.rda' %
                                                      (_('All importable data files'),
                                                       _('Spreadsheet files'), _('Text files'),
                                                       _('SPSS data files'), _('SAS data files'), _('STATA data files'),
@@ -44,10 +44,10 @@ def open_demo_data_file():
                                                      '/sample_data',
                                                      '%s (*.ods *.xls *.xlsx *.csv *.txt *.tsv *.dat *.log '
                                                      '*.sav *.zsav *.por *.sas7bdat *.xpt *.dta '
-                                                     '*.rdata *.Rdata *.rds);;'
+                                                     '*.rdata *.Rdata *.rds *.rda);;'
                                                      '%s *.ods *.xls *xlsx;;%s *.csv *.txt *.tsv *.dat *.log;;'
                                                      '%s *.sav *.zsav *.por;;%s *.sas7bdat *.xpt;; %s *.dta;;'
-                                                     '%s *.rdata *.Rdata *.rds' %
+                                                     '%s *.rdata *.Rdata *.rds *.rda' %
                                                      (_('All importable data files'),
                                                       _('Spreadsheet files'), _('Text files'),
                                                       _('SPSS data files'), _('SAS data files'), _('STATA data files'),