+++ changelog.md
@@ -9,7 +9,7 @@ Upcoming release
     - Excel spreadsheet .xls and xlsx files
     - OpenDocument Spreadsheet .ods files
     - SPSS .zsav and .por files
-    - R .rdata and .rds files
+    - R .rdata, .rds, and .rda files
     - SAS .sas7bdat and .xpt files
     - STATA .dta files
     - Additional .csv formats